requirements.txt
azure-cosmos
azure-storage-blob
gradio
python-dotenv
numpy
transformers
torch
ultralytics
azure-functions
